+++ Engine.vcxproj
@@ -544,8 +544,8 @@
     <ClInclude Include="include\DebugUI.hpp" />
     <ClInclude Include="include\IGame.hpp" />
     <ClInclude Include="include\Utils.hpp" />
-    <ClInclude Include="src\DirectX\DirectXAdapter.hpp" />
-    <ClInclude Include="src\DirectX\Heap\Heap.hpp" />
+    <ClInclude Include="src\Renderer\DirectX\DirectXAdapter.hpp" />
+    <ClInclude Include="src\Renderer\DirectX\Heap\Heap.hpp" />
     <ClInclude Include="src\Engine\Config\Config.hpp" />
     <ClInclude Include="src\Engine\Engine.hpp" />
     <ClInclude Include="include\Log.hpp" />
@@ -553,6 +553,7 @@
     <ClInclude Include="include\Input.hpp" />
     <ClInclude Include="include\Math\Vector4.hpp" />
     <ClInclude Include="src\Platform\WinApp.hpp" />
+    <ClInclude Include="src\Renderer\Renderer.hpp" />
     <ClInclude Include="src\Scheduler\Scheduler.hpp" />
     <ClInclude Include="include\Singleton.hpp" />
     <ClInclude Include="src\Timer\Timer.hpp" />
@@ -560,8 +561,8 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Debug\DebugUI.cpp" />
-    <ClCompile Include="src\DirectX\DirectXAdapter.cpp" />
-    <ClCompile Include="src\DirectX\Heap\Heap.cpp" />
+    <ClCompile Include="src\Renderer\DirectX\DirectXAdapter.cpp" />
+    <ClCompile Include="src\Renderer\DirectX\Heap\Heap.cpp" />
     <ClCompile Include="src\Engine\Config\Config.cpp" />
     <ClCompile Include="src\Engine\Engine.cpp" />
     <ClCompile Include="src\Framework\Framework.cpp" />
@@ -569,6 +570,7 @@
     <ClCompile Include="src\Input\Input.cpp" />
     <ClCompile Include="src\Math\Vector4.cpp" />
     <ClCompile Include="src\Platform\WinApp.cpp" />
+    <ClCompile Include="src\Renderer\Renderer.cpp" />
     <ClCompile Include="src\Scheduler\Scheduler.cpp" />
     <ClCompile Include="src\System\Log.cpp" />
     <ClCompile Include="src\System\Singleton\Singleton.cpp" />
