+++ Engine.vcxproj.filters
@@ -45,7 +45,7 @@
     <ClInclude Include="include\Singleton.hpp">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="src\DirectX\DirectXAdapter.hpp">
+    <ClInclude Include="src\Renderer\DirectX\DirectXAdapter.hpp">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
     <ClInclude Include="src\Engine\Config\Config.hpp">
@@ -60,7 +60,10 @@
     <ClInclude Include="include\DebugUI.hpp">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="src\DirectX\Heap\Heap.hpp">
+    <ClInclude Include="src\Renderer\DirectX\Heap\Heap.hpp">
+      <Filter>ヘッダー ファイル</Filter>
+    </ClInclude>
+    <ClInclude Include="src\Renderer\Renderer.hpp">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
   </ItemGroup>
@@ -95,7 +98,7 @@
     <ClCompile Include="src\System\Singleton\Singleton.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="src\DirectX\DirectXAdapter.cpp">
+    <ClCompile Include="src\Renderer\DirectX\DirectXAdapter.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
     <ClCompile Include="src\Engine\Config\Config.cpp">
@@ -110,7 +113,10 @@
     <ClCompile Include="src\Debug\DebugUI.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="src\DirectX\Heap\Heap.cpp">
+    <ClCompile Include="src\Renderer\DirectX\Heap\Heap.cpp">
+      <Filter>ソース ファイル</Filter>
+    </ClCompile>
+    <ClCompile Include="src\Renderer\Renderer.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
   </ItemGroup>
